+++ .dockerignore
@@ -1,2 +1,3 @@
 **
-!release
+!release-github
+!ci
