+++ .travis.yml
@@ -1,28 +1,33 @@
-language: minimal
+language: node_js
+node_js: node
 
 jobs:
   include:
     - name: Test
       if: tag IS blank
-      script: ./ci/image/run.sh "yarn && git submodule update --init && yarn vscode:patch && ./ci/ci.sh"
+      script: ./ci/container/exec.sh ./ci/steps/test.sh
       deploy: null
+      install: null
     - name: Linux Release
       if: tag IS present
-      script:
-        - travis_wait 60 ./ci/image/run.sh "yarn && yarn vscode && ci/release.sh && ./ci/build-test.sh"
-        - ./ci/release-image/push.sh
-    - name: Linux ARM64 Release
+      script: ./ci/steps/linux-release.sh
+      install: null
+    - name: Linux Release
       if: tag IS present
-      script:
-        - ./ci/image/run.sh "yarn && yarn vscode && ci/release.sh && ./ci/build-test.sh"
-        - ./ci/release-image/push.sh
       arch: arm64
+      script: |
+        sudo apt-get update && sudo apt-get install -y jq || exit 1
+        ./ci/steps/linux-release.sh
+      install: null
     - name: MacOS Release
       if: tag IS present
       os: osx
-      language: node_js
+      # node 13/14 crashes in the build process for some reason.
       node_js: 12
-      script: yarn && yarn vscode && travis_wait 60 ci/release.sh && ./ci/build-test.sh
+      script: |
+        HOMEBREW_NO_INSTALL_CLEANUP=1 HOMEBREW_NO_AUTO_UPDATE=1 brew install jq || exit 1
+        travis_wait 60 ./ci/steps/static-release.sh || exit 1
+      install: null
 
 before_deploy:
   - echo "$JSON_KEY" | base64 --decode > ./ci/key.json
@@ -36,8 +41,10 @@ deploy:
     target_commitish: $TRAVIS_COMMIT
     name: $TRAVIS_TAG
     file:
-      - release/*.tar.gz
-      - release/*.zip
+      - release-github/*.tar.gz
+      - release-github/*.zip
+      - release-github/*.deb
+      - release-github/*.rpm
     on:
       tags: true
   - provider: gcs
@@ -45,14 +52,25 @@ deploy:
     bucket: "codesrv-ci.cdr.sh"
     upload_dir: "releases"
     key_file: ./ci/key.json
-    local_dir: release-upload
+    local_dir: ./release-gcp
     on:
       tags: true
       # TODO: The gcs provider fails to install on arm64.
-      condition: $TRAVIS_CPU_ARCH = amd64
+      condition: $TRAVIS_CPU_ARCH == amd64
+  - provider: script
+    edge: true
+    # We do not use the travis npm deploy integration as it does not allow us to
+    # deploy a subpath and and v2 which should, just errors out that the src does not exist
+    script: ./ci/steps/publish-npm.sh
+    on:
+      tags: true
+      condition: $TRAVIS_CPU_ARCH == amd64 && $TRAVIS_OS_NAME == linux
 
 cache:
   timeout: 600
   yarn: true
   directories:
+    - .cache
+    - out
+    - dist
     - lib/vscode/.build/extensions
