+++ ci/dev/fmt.sh
@@ -1,8 +1,9 @@
 #!/usr/bin/env bash
-
 set -euo pipefail
 
 main() {
+  cd "$(dirname "$0")/../.."
+
   shfmt -i 2 -w -s -sr $(git ls-files "*.sh")
 
   local prettierExts
