+++ ci/release-container/push.sh
@@ -5,18 +5,21 @@ set -euo pipefail
 main() {
   cd "$(dirname "$0")/../.."
   source ./ci/lib.sh
-  set_version
+  VERSION="$(pkg_json_version)"
 
-  if [[ ${CI:-} ]]; then
+  if [[ ${CI-} ]]; then
     echo "$DOCKER_PASSWORD" | docker login -u "$DOCKER_USERNAME" --password-stdin
   fi
 
   imageTag="codercom/code-server:$VERSION"
-  if [[ ${TRAVIS_CPU_ARCH:-} == "arm64" ]]; then
+  if [[ $(arch) == "arm64" ]]; then
     imageTag+="-arm64"
   fi
-  docker build -t "$imageTag" -f ./ci/release-image/Dockerfile .
-  docker push codercom/code-server
+
+  docker build \
+    -t "$imageTag" \
+    -f ./ci/release-container/Dockerfile .
+  docker push "$imageTag"
 }
 
 main "$@"
